Remove version munging
Description: We found that when building wheels for vllm it would build something akin to the following: vllm-0.5.2+cu124-cp311-cp311-linux_x86_64.whl.
From this, twine thinks the package version is not 0.5.2, but instead 0.5.2+cu124, which causes resolution and upload issues. We want both the
patched Sdist and the wheel to live in the same package, just with different files. This change should remove the `+cu124` we were seeing for cuda, but still
allow for checking for unknown build-environment errors.

Co-Authored-By: Greg Pereira <grpereir@redhat.com>
Signed-off-by: Christian Heimes <christian@python.org>